整理后台页面，修复后端部分代码结构问题

+++ RuoYi-Vue/lee-site/src/main/java/com/ruoyi/lee/enums/LeeProjectStatusEnums.java
@@ -21,13 +21,13 @@ public enum LeeProjectStatusEnums {
     ;
 
 
-    private final int status;
+    private final long status;
 
     LeeProjectStatusEnums(int status) {
         this.status = status;
     }
 
-    public int getStatus() {
+    public long getStatus() {
         return status;
     }
 }
